+++ src/entity/static_mesh.hpp
@@ -1,5 +1,5 @@
-#include "mesh.hpp"
 #include "../gui/meshloader.hpp"
+#include "mesh.hpp"
 
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
@@ -46,14 +46,14 @@ class StaticMesh : public MeshLoader::MultiComplexMesh {
 
     inline StaticMesh(){};
 
-    StaticMesh(entity::Mesh mesh, const std::vector<glm::vec3>& model_transforms);
+    StaticMesh(entity::Mesh mesh, const std::vector<glm::ivec3>& model_transforms);
 
     StaticMesh(
         const std::vector<unsigned short>& indices,
-        const std::vector<glm::vec3>& indexed_vertices,
+        const std::vector<glm::ivec3>& indexed_vertices,
         const std::vector<glm::vec3>& indexed_colors,
-        const std::vector<glm::vec3>& indexed_normals,
-        const std::vector<glm::vec3>& model_transforms
+        const std::vector<glm::ivec3>& indexed_normals,
+        const std::vector<glm::ivec3>& model_transforms
     );
 
     inline ~StaticMesh() {
