+++ src/entity/terrain_mesh.hpp
@@ -42,17 +42,17 @@ class TerrainMesh : public MeshLoader::SingleComplexMesh {
     TerrainMesh(entity::Mesh mesh);
     void init(
         const std::vector<unsigned short>& indices,
-        const std::vector<glm::vec3>& indexed_vertices,
+        const std::vector<glm::ivec3>& indexed_vertices,
         const std::vector<glm::vec3>& indexed_colors,
-        const std::vector<glm::vec3>& indexed_normals
+        const std::vector<glm::ivec3>& indexed_normals
     );
     void init(entity::Mesh mesh);
 
     TerrainMesh(
         const std::vector<unsigned short>& indices,
-        const std::vector<glm::vec3>& indexed_vertices,
+        const std::vector<glm::ivec3>& indexed_vertices,
         const std::vector<glm::vec3>& indexed_colors,
-        const std::vector<glm::vec3>& indexed_normals
+        const std::vector<glm::ivec3>& indexed_normals
     );
 
     inline ~TerrainMesh() {
